move events to its own file, now each csproj has a EventSourceEvents.cs
