+++ src/Orleans.Core/Runtime/CallbackData.cs
@@ -3,42 +3,11 @@ using System.Diagnostics.Tracing;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Orleans.DistributedTracing.EventSourceEvents;
 using Orleans.Transactions;
 
 namespace Orleans.Runtime
 {
-    [EventSource(Name = "Microsoft-Orleans-CallBackDataEvent")]
-    public class OrleansCallBackDataEvent : EventSource
-    {
-        public static readonly OrleansCallBackDataEvent Log = new OrleansCallBackDataEvent();
-        public void OnTimeoutStart()
-        {
-            WriteEvent(1);
-        }
-        public void OnTimeoutStop()
-        {
-            WriteEvent(2);
-        }
-
-        public void OnTargetSiloFailStart()
-        {
-            WriteEvent(3);
-        }
-        public void OnTargetSiloFailStop()
-        {
-            WriteEvent(4);
-        }
-
-        public void DoCallbackStart()
-        {
-            WriteEvent(5);
-        }
-        public void DoCallbackStop()
-        {
-            WriteEvent(6);
-        }
-    }
-
     internal class CallbackData
     {
         private readonly SharedCallbackData shared;
